more small refactorings
git-svn-id: https://svn.codehaus.org/groovy/trunk/groovy/modules/native_launcher@15355 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ source/jst_fileutils.c
@@ -786,7 +786,7 @@ static int stripExecNameAndLastDirFromPath( char* executablePath, const char* la
     if ( lastDirLen ) {
       size_t len = strlen( executablePath ) ;
 
-      if ( ( (int)len >= lastDirLen + 1 ) &&
+      if ( ( len >= lastDirLen + 1 ) &&
            jst_endsWith( executablePath, lastDirOnExecPath ) ) {
 
         found = JNI_TRUE ;
